+++ apps/cli/src/helpers/project-generation/create-project.ts
@@ -13,6 +13,7 @@ import {
 	generateCloudflareWorkerTypes,
 	setupRuntime,
 } from "../setup/runtime-setup";
+import { setupWebDeploy } from "../setup/web-deploy-setup";
 import { createReadme } from "./create-readme";
 import { setupEnvironmentVariables } from "./env-setup";
 import { initializeGit } from "./git";
@@ -26,6 +27,8 @@ import {
 	setupAuthTemplate,
 	setupBackendFramework,
 	setupDbOrmTemplates,
+	setupDeploymentTemplates,
+	setupDockerComposeTemplates,
 	setupExamplesTemplate,
 	setupFrontendTemplates,
 } from "./template-manager";
@@ -42,12 +45,15 @@ export async function createProject(options: ProjectConfig) {
 		await setupBackendFramework(projectDir, options);
 		if (!isConvex) {
 			await setupDbOrmTemplates(projectDir, options);
+			await setupDockerComposeTemplates(projectDir, options);
 			await setupAuthTemplate(projectDir, options);
 		}
 		if (options.examples.length > 0 && options.examples[0] !== "none") {
 			await setupExamplesTemplate(projectDir, options);
 		}
 		await setupAddonsTemplate(projectDir, options);
+
+		await setupDeploymentTemplates(projectDir, options);
 
 		await setupApi(options);
 
@@ -70,6 +76,8 @@ export async function createProject(options: ProjectConfig) {
 
 		await handleExtras(projectDir, options);
 
+		await setupWebDeploy(options);
+
 		await setupEnvironmentVariables(options);
 		await updatePackageConfigurations(projectDir, options);
 		await createReadme(projectDir, options);
@@ -88,7 +96,7 @@ export async function createProject(options: ProjectConfig) {
 
 		await initializeGit(projectDir, options.git);
 
-		displayPostInstallInstructions({
+		await displayPostInstallInstructions({
 			...options,
 			depsInstalled: options.install,
 		});