+++ apps/cli/src/types.ts
@@ -1,4 +1,4 @@
-import { z } from "zod";
+import z from "zod";
 
 export const DatabaseSchema = z
 	.enum(["none", "sqlite", "postgres", "mysql", "mongodb"])
@@ -61,6 +61,7 @@ export const DatabaseSetupSchema = z
 		"mongodb-atlas",
 		"supabase",
 		"d1",
+		"docker",
 		"none",
 	])
 	.describe("Database hosting setup");
@@ -92,6 +93,11 @@ export const ProjectNameSchema = z
 	.describe("Project name or path");
 export type ProjectName = z.infer<typeof ProjectNameSchema>;
 
+export const WebDeploySchema = z
+	.enum(["workers", "none"])
+	.describe("Web deployment");
+export type WebDeploy = z.infer<typeof WebDeploySchema>;
+
 export type CreateInput = {
 	projectName?: string;
 	yes?: boolean;
@@ -108,10 +114,12 @@ export type CreateInput = {
 	backend?: Backend;
 	runtime?: Runtime;
 	api?: API;
+	webDeploy?: WebDeploy;
 };
 
 export type AddInput = {
 	addons?: Addons[];
+	webDeploy?: WebDeploy;
 	projectDir?: string;
 	install?: boolean;
 	packageManager?: PackageManager;
@@ -138,6 +146,7 @@ export interface ProjectConfig {
 	install: boolean;
 	dbSetup: DatabaseSetup;
 	api: API;
+	webDeploy: WebDeploy;
 }
 
 export interface BetterTStackConfig {
@@ -154,6 +163,7 @@ export interface BetterTStackConfig {
 	packageManager: PackageManager;
 	dbSetup: DatabaseSetup;
 	api: API;
+	webDeploy: WebDeploy;
 }
 
 export type AvailablePackageManagers = "npm" | "pnpm" | "bun";